Refactoring the execution Context class to support CQL Population contexts and provide additional support for included libraries in CQL documents when evaluating CQL Patient context statements.
The execution engines Context class has been subclassed into PopulationContext and PatientContext.  The PatientContext class is used as the root context when evaluating individual patients.  As the PatientContext is evaluated for a single patient this allows for included libraries to cache execution results on a patient basis in the PatientContext eliminating the need to re-evaluate the statements if called upon more then once in a document.

The PopulationContext acts as the root context for evaluating a set of patient level results.  When a CQL Population context statement is evaluated any CQL Patient level expressions that it requests will be represented as an array of the Patient level results for that statement.  As an example consider the following contrived example.

parameter MeasurementPeriod default Interval[DateTime(2013, 1, 1), DateTime(2014, 1, 1))

context Patient

define Age:
  AgeInYearsAt(start of MeasurementPeriod)

context Population

define age_sum: Sum(Age)

For every patient that is evaluated against the CQL Patient context above an Age entry will be calculated and stored in a set of results.

{Patient_id_1: {Age: 20},
 Patient_id_2: {Age: 10},
 Patient_id_3: {Age: 30}
}

When evaluating the Population context statement  age_sum we see that there is a reference to the Age statement from the Patient context.  For the purposes for the PopulatioNContext class when the Age statement is referenced it will loop over the set of patient level results and pull out the value for the Age statement that is stored for each of the patients and return that as an array of the values.  ex. [20,10,30] .  From there the Sum function will operate as it normally does and simply perform a summation of the values in the array (60) and then the PopulationContext will store that as the value for the age_sum statement.

Included in this refactor is an Executor class.  The executor class is an abstraction of the execution that was being performed in the Library class.