Fix Null Detection and Null Variable Retrieval
Previously, if undefined was executed against IsNull it would return false.  Not it returns true.

Also fixed an issue in Context.get.  Previously, if you did a get on a variable, and it *was* in the context with a value of null, the context would instead look for the variable in the parent context.  Now it only looks at the parent if the variable is *undefined* in the current context.  Variables that are literal null should be returned as null.