Fix Incorrect Use of Should
In many tests, we asserted `foo.should.be.true` and `foo.should.be.false`.  As it turns out, these assertions *never* fail because they are incorrectly specified.  The correct usage is `foo.should.be.true()` and `foo.should.be.false()`.  This commit fixes that (in a ridiculous number of places).