CQL Exec: Rename DurationBetween to DifferenceBetween
Early versions of CQL (prior to 1.0) defined DurationBetween to count boundaries between.  In the 1.0 release, this was changed to count elapsed time, and DifferenceBetween was introduced to count boundaries.

Unfortunately, the CQL execution engine still had DurationBetween implemented against the boundary-counting definition.  This is incorrect and has been fixed by this commit, which gives it the proper name: DifferenceBetween.

A future commit will implement the real Duration Between.