Remove version check for equivalency operator.
Affects Code, Value Set, and Concept equivalency.
Equivalency should no longer return false if the only difference is in version.

Modified Equivalent class to use the equivalent function in comparisn.js rather
than a different but similar expression.

The only tests that failed due to this change were expected (tests that were
looking for different versions). I updated those tests to expect equivalence if
the only difference is in version.