In CqlEngine, prevent windows from accumulating for multiple evaluations
Each call to CqlEngine.evaluate causes a call to State.init which,
before this change, pushed one window. The actual evaluation
associated with the CqlEngine.evaluate call pushed and popped a
balanced number of windows but the one pushed by State.init remained
on the stack in the end. Successive CqlEngine.evaluate calls would
therefore accumulate an unbounded number of windows on the stack.

As a slight interface change, clients that bypass
CqlEngine.evaluate[Expressions] by using
CqlEngine.getEvaluationVisitor now have to call State.beginEvaluation
and State.endEvaluation themselves. The
ExpressionCacheTest.expression_cache test had to be change in this way.