Minor improvements and possible change for ToString(Datetime) (#1551)
* Handle return value of Collection.binarySearch() correctly

If the element is not found, the return value is not constantly -1 but
instead an encoding of the position at which the element should be
inserted.

* When reporting invalid Interval bounds, include offending bounds

* For DateTime, include timezone offset in printed representations

This change affects two printed representations of the DateTime class:

1. On the Java level, DateTime.toString() now includes the timezone
   offset

2. On the CQL level, ToString(<DateTime>) now includes the timezone
   offset as well

Of these two, 2. is clearly the important change since it affects CQL
semantics. The justification is in the specification of
ToString (https://cql.hl7.org/09-b-cqlreference.html#tostring):

1. In the table which lists the expected output shape of ToString for
   different types, the row for DateTime contains

     YYYY-MM-DDThh:mm:ss.fff(+|-)hh:mm

   in which (+|-)hh:mm must represent the timezone offset.

2. The requirement "The result of any ToString must be round-trippable
   back to the source value." would not be satisfied if the printed
   representation did not include the timezone offset.

* In CQLCompiler, strip double quotes form library identifier in errors

Before this change, the library identifier in thrown syntax errors
included delimiting double quotes.

In other thrown errors, the library identifier does not include double
quotes. Also, the double quotes are part of the syntax, not the actual
library name so stripping seems like the better behavior regardless.

* Various minor simplifications

* Fix typos in comments

---------

Co-authored-by: JP <jonathan.i.percival@gmail.com>