Add support for Backtraces and profiling (#1550)
* In CqlEngine, prevent windows from accumulating for multiple evaluations

Each call to CqlEngine.evaluate causes a call to State.init which,
before this change, pushed one window. The actual evaluation
associated with the CqlEngine.evaluate call pushed and popped a
balanced number of windows but the one pushed by State.init remained
on the stack in the end. Successive CqlEngine.evaluate calls would
therefore accumulate an unbounded number of windows on the stack.

As a slight interface change, clients that bypass
CqlEngine.evaluate[Expressions] by using
CqlEngine.getEvaluationVisitor now have to call State.beginEvaluation
and State.endEvaluation themselves. The
ExpressionCacheTest.expression_cache test had to be change in this way.

* In ExpressionRefEvaluator, push a window for local variables

This makes the association between local variables introduced by, say,
query expressions and the containing expression definition clear. For
function definitions, there is no change since function invocation
already pushed a fresh window.

* Use a stack of ActivationFrames instead of the window stack

The stack of ActivationFrames serves the same purpose as the window
stack but records additional information compared to just groups of
local variables:

* Each ActivationFrame (except the dummy root frame) has an associated
  element which is an ExpressionDef (possibly a FunctionDef) or a
  Retrieve.

* Each ActivationFrame has a start time and an end time so that the
  time spent evaluating the frame can be tracked.

* Each ActivationFrame has an isCached field so that the cache hits
  and misses when evaluating ExpressionDef elements (but not
  FunctionDef elements) can be tracked.

This additional information will be useful for at least two purposes:

1. The associated elements can be used to construct a backtrace with
   source information when an exception is thrown during evaluation.

2. The timing and cache hit information can be used to build a
   deterministic profile of a given evaluation which is useful for
   performance analysis.

Subsequent commits will add code for these two use-cases.

* Include a Backtrace (new class) in CqlExceptions thrown at runtime

A backtrace represents a stack of call-like frames from the root of an
evaluation to a particular sub-expression, commonly a sub-expression
in which an exception was thrown. Client code can extract this
Backtrace object from a caught exception and use it to display a
backtrace similar to those known from other programming languages.

* New Profile class

* Track context in ActivationFrames

* Try to produce better colors in flamegraph

* Include context information in backtrace frames

* Remove most profiling overhead

Before this change, there were two performance issues in the profile
construction code

1. Profile.register traversed the whole stack to find the appropriate
   profile node which is O(N) where N is the depth of the "call stack".

2. Children of profile nodes were stored in a HashMap the keys of
   which relied on object-equality involving Element
   objects. Object-equality for those is very expensive.

This change solves issue 1 by remembering the stack of profile nodes
that correspond to the stack of current activation frames and
incrementally updating the top of the stack which is O(1).

This change solves issue 2 by (basically; the context adds a slight
complication) storing the children of profile nodes in an
IdentityHashMap so that Element.equals is not used.

* Add method Variable.toString

* Add method Profile.Node.toString

* In Backtrace.Frame, order of arguments matches function signature

* Improve documentation of Profile class

* Improve documentation of Backtrace class

* Initial test for Backtrace functionality

* Initial test for profiling functionality

* Use multiple different Engine Options in CqlPerformanceIT

* Remove unused code in ProfilingTest

* Fix typos in comments

Reported-by: Anton Vasetenkov <antvaset@gmail.com>

* Use exceptions instead of assertions for stack underflow in State

Suggested-by: Anton Vasetenkov <antvaset@gmail.com>

* Fix merge/rebase damage in State

Suggested-by: Anton Vasetenkov <antvaset@gmail.com>

* Simplify activation frame creation FunctionRefEvaluator

Suggested-by: Anton Vasetenkov <antvaset@gmail.com>

* Use pattern matching instanceof since language level now permits it

The language level was changed in commit
b55dc7bd5f4fae27d507120321a011831820d910.

Suggested-by: Anton Vasetenkov <antvaset@gmail.com>

---------

Co-authored-by: JP <jonathan.i.percival@gmail.com>
Co-authored-by: c-schuler <hoofschu@gmail.com>
Co-authored-by: Anton Vasetenkov <antvaset@gmail.com>